[Tests] Split test into multiple files
