+++ pubspec.yaml
@@ -24,6 +24,7 @@ dependencies:
   darq: ^1.2.1
   device_info_plus: ^4.0.0
   devicelocale: ^0.5.4
+  dio: ^4.0.6
   enum_to_string: ^2.0.1
   fl_chart: ^0.55.0
   flutter:
@@ -64,6 +65,7 @@ dependencies:
   screenshot: ^1.2.3
   shared_preferences: ^2.0.15
   sprintf: ^6.0.0
+  store_checker: ^1.0.0
   synchronized: ^3.0.0+2
   transparent_image: 2.0.0
   tuple: ^2.0.0
