+++ test/application/custom_builds/custom_builds_bloc_test.dart
@@ -15,6 +15,7 @@ const _dbFolder = 'shiori_custom_builds_bloc_tests';
 void main() {
   late DataService _dataService;
   late GenshinService _genshinService;
+  late final String _dbPath;
 
   const _keqingKey = 'keqing';
   const _ganyuKey = 'ganyu';
@@ -30,14 +31,15 @@ void main() {
 
     return Future(() async {
       await _genshinService.init(AppLanguageType.english);
-      await _dataService.init(dir: _dbFolder);
+      _dbPath = await getDbPath(_dbFolder);
+      await _dataService.initForTests(_dbPath);
     });
   });
 
   tearDownAll(() {
     return Future(() async {
       await _dataService.closeThemAll();
-      await deleteDbFolder(_dbFolder);
+      await deleteDbFolder(_dbPath);
     });
   });
