+++ test/application/notification/notification_bloc_test.dart
@@ -26,6 +26,7 @@ void main() {
   late final GenshinService _genshinService;
   late final DataService _dataService;
   late final NotificationsBloc _notificationsBloc;
+  late final String _dbPath;
 
   const _defaultTitle = 'Notification title';
   const _defaultBody = 'Notification body';
@@ -53,14 +54,15 @@ void main() {
 
     return Future(() async {
       await _genshinService.init(_settingsService.language);
-      await _dataService.init(dir: _dbFolder);
+      _dbPath = await getDbPath(_dbFolder);
+      await _dataService.initForTests(_dbPath);
     });
   });
 
   tearDownAll(() {
     return Future(() async {
       await _dataService.closeThemAll();
-      await deleteDbFolder(_dbFolder);
+      await deleteDbFolder(_dbPath);
     });
   });
