+++ test/application/notifications/notifications_bloc_test.dart
@@ -18,6 +18,7 @@ void main() {
   late final SettingsService _settingsService;
   late final MockNotificationService _notificationService;
   late final DataService _dataService;
+  late final String _dbPath;
 
   const _defaultTitle = 'Notification title';
   const _defaultBody = 'Notification body';
@@ -44,14 +45,15 @@ void main() {
 
     return Future(() async {
       await genshinService.init(_settingsService.language);
-      await _dataService.init(dir: _dbFolder);
+      _dbPath = await getDbPath(_dbFolder);
+      await _dataService.initForTests(_dbPath);
     });
   });
 
   tearDownAll(() {
     return Future(() async {
       await _dataService.closeThemAll();
-      await deleteDbFolder(_dbFolder);
+      await deleteDbFolder(_dbPath);
     });
   });
