+++ test/application/tier_list/tier_list_bloc_test.dart
@@ -19,6 +19,7 @@ void main() {
   late final LoggingService _loggingService;
   late final GenshinService _genshinService;
   late final DataService _dataService;
+  late final String _dbPath;
 
   setUpAll(() {
     TestWidgetsFlutterBinding.ensureInitialized();
@@ -30,14 +31,15 @@ void main() {
 
     return Future(() async {
       await _genshinService.init(AppLanguageType.english);
-      await _dataService.init(dir: _dbFolder);
+      _dbPath = await getDbPath(_dbFolder);
+      await _dataService.initForTests(_dbPath);
     });
   });
 
   tearDownAll(() {
     return Future(() async {
       await _dataService.closeThemAll();
-      await deleteDbFolder(_dbFolder);
+      await deleteDbFolder(_dbPath);
     });
   });
