+++ test/application/weapon/weapon_bloc_test.dart
@@ -16,6 +16,7 @@ void main() {
   late final TelemetryService _telemetryService;
   late final GenshinService _genshinService;
   late final DataService _dataService;
+  late final String _dbPath;
 
   const _key = 'aquila-favonia';
 
@@ -29,14 +30,15 @@ void main() {
 
     return Future(() async {
       await _genshinService.init(AppLanguageType.english);
-      await _dataService.init(dir: _dbFolder);
+      _dbPath = await getDbPath(_dbFolder);
+      await _dataService.initForTests(_dbPath);
     });
   });
 
   tearDownAll(() {
     return Future(() async {
       await _dataService.closeThemAll();
-      await deleteDbFolder(_dbFolder);
+      await deleteDbFolder(_dbPath);
     });
   });
