+++ test/common.dart
@@ -3,19 +3,26 @@ import 'dart:io';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:intl/date_symbol_data_local.dart';
-import 'package:path/path.dart' as path_helper;
-import 'package:path_provider/path_provider.dart';
+import 'package:mockito/mockito.dart';
 import 'package:shiori/domain/enums/enums.dart';
 import 'package:shiori/domain/extensions/string_extensions.dart';
 import 'package:shiori/domain/models/models.dart';
-import 'package:shiori/domain/services/genshin_service.dart';
+import 'package:shiori/domain/services/file/file_infrastructure.dart';
 import 'package:shiori/domain/services/locale_service.dart';
+import 'package:shiori/infrastructure/infrastructure.dart';
+
+import 'mocks.mocks.dart';
+
+Future<String> getDbPath(String subDir) async {
+  TestWidgetsFlutterBinding.ensureInitialized();
+  final appDir = await Directory.systemTemp.createTemp(subDir);
+  await Directory(appDir.path).create();
+  return appDir.path;
+}
 
 //since we are using a real impl of the data service,
 // to avoid problems we just create different folders and delete them all after the test completes
-Future<void> deleteDbFolder(String subDir) async {
-  final appDir = await getApplicationDocumentsDirectory();
-  final path = path_helper.join(appDir.path, subDir);
+Future<void> deleteDbFolder(String path) async {
   await Directory(path).delete(recursive: true);
 }
 
@@ -46,9 +53,21 @@ Future<bool> _assetExists(String path) async {
   }
 }
 
-void checkAsset(String path) {
+Future<bool> _fileExists(String path) => File(path).exists();
+
+void checkAsset(String path, {bool isAnAsset = true}) {
   expect(path, allOf([isNotEmpty, isNotNull]));
-  expect(_assetExists(path), completion(equals(true)), reason: 'Asset = $path does not exist');
+  if (isAnAsset) {
+    expect(_assetExists(path), completion(equals(true)), reason: 'Asset = $path does not exist');
+  } else {
+    expect(_fileExists(path), completion(equals(true)), reason: 'File = $path does not exist');
+  }
+}
+
+void checkAssets(List<String> paths, {bool isAnAsset = false}) {
+  for (final path in paths) {
+    checkAsset(path, isAnAsset: isAnAsset);
+  }
 }
 
 void checkItemsCommon(List<ItemCommon> items, {bool checkEmpty = true}) {
@@ -80,29 +99,29 @@ void checkItemKeyNameAndImage(String key, String name, String image) {
   checkTranslation(name, canBeNull: false);
 }
 
-void checkItemAscensionMaterialFileModel(GenshinService service, List<ItemAscensionMaterialFileModel> all) {
+void checkItemAscensionMaterialFileModel(MaterialFileService materialFileService, List<ItemAscensionMaterialFileModel> all) {
   expect(all, isNotEmpty);
   for (final material in all) {
     checkKey(material.key);
-    expect(() => service.materials.getMaterial(material.key), returnsNormally);
+    expect(() => materialFileService.getMaterial(material.key), returnsNormally);
     expect(material.quantity, greaterThanOrEqualTo(0));
   }
 }
 
-void checkCharacterFileAscensionMaterialModel(GenshinService service, List<CharacterFileAscensionMaterialModel> all) {
+void checkCharacterFileAscensionMaterialModel(MaterialFileService materialFileService, List<CharacterFileAscensionMaterialModel> all) {
   expect(all, isNotEmpty);
   for (final ascMaterial in all) {
     expect(ascMaterial.rank, allOf([greaterThanOrEqualTo(1), lessThanOrEqualTo(6)]));
     expect(ascMaterial.level, allOf([greaterThanOrEqualTo(20), lessThanOrEqualTo(80)]));
-    checkItemAscensionMaterialFileModel(service, ascMaterial.materials);
+    checkItemAscensionMaterialFileModel(materialFileService, ascMaterial.materials);
   }
 }
 
-void checkCharacterFileTalentAscensionMaterialModel(GenshinService service, List<CharacterFileTalentAscensionMaterialModel> all) {
+void checkCharacterFileTalentAscensionMaterialModel(MaterialFileService materialFileService, List<CharacterFileTalentAscensionMaterialModel> all) {
   expect(all, isNotEmpty);
   for (final ascMaterial in all) {
     expect(ascMaterial.level, inInclusiveRange(2, 10));
-    checkItemAscensionMaterialFileModel(service, ascMaterial.materials);
+    checkItemAscensionMaterialFileModel(materialFileService, ascMaterial.materials);
   }
 }
 
@@ -119,4 +138,12 @@ void checkTranslation(String? text, {bool canBeNull = true, bool checkForColor =
     final hasColor = text.contains('{color}') || text.contains('{/color}');
     expect(hasColor, isFalse);
   }
+}
+
+LocaleService getLocaleService(AppLanguageType language) {
+  final settings = MockSettingsService();
+  when(settings.language).thenReturn(language);
+  final service = LocaleServiceImpl(settings);
+  manuallyInitLocale(service, language);
+  return service;
 }
