+++ test/infrastructure/data_service_test.dart
@@ -18,6 +18,7 @@ void main() {
   late final DataService _dataService;
   late final CalculatorService _calculatorService;
   late final GenshinService _genshinService;
+  late final String _dbPath;
 
   setUp(() {
     TestWidgetsFlutterBinding.ensureInitialized();
@@ -31,14 +32,15 @@ void main() {
 
     return Future(() async {
       await _genshinService.init(AppLanguageType.english);
-      await _dataService.init(dir: _dbFolder);
+      _dbPath = await getDbPath(_dbFolder);
+      await _dataService.initForTests(_dbPath);
     });
   });
 
   tearDown(() {
     return Future(() async {
       await _dataService.closeThemAll();
-      await deleteDbFolder(_dbFolder);
+      await deleteDbFolder(_dbPath);
     });
   });
